Extension for netprop list
I've started working on a small extension that allows pawn to
iterate over a list of all netprops.
There is no documentation yet, but an example which should be
pretty easy to read. Its output should be the same as
sm_dump_netprops.
I will not provide builds for this. But for the examples sake
i've included linux binaries for l4d2 and tf2.